+++ app/javascript/vue/components/radials/object/radial.vue
@@ -104,7 +104,8 @@ import Icons from './images/icons.js'
 import { useShortcuts } from '@/components/radials/composables'
 import { SLICE, SLICES_BY_OBJECT_TYPE } from './constants/slices.js'
 import { Tag } from '@/routes/endpoints'
-import { ref, computed, onMounted, getCurrentInstance } from 'vue'
+import { ref, computed, onMounted } from 'vue'
+import { randomUUID } from '@/helpers'
 
 const MIDDLE_RADIAL_BUTTON = 'circleButton'
 
@@ -165,7 +166,7 @@ const isVisible = ref(false)
 const metadata = ref(null)
 const title = ref('Quick forms')
 const defaultTag = ref(null)
-const headerRightTargetId = `radial-object-header-right-${getCurrentInstance().uid}`
+const headerRightTargetId = `radial-object-header-right-${randomUUID()}`
 const { removeListener, setShortcutsEvent } = useShortcuts({
   metadata,
   currentAnnotator