+++ app/javascript/vue/tasks/collection_objects/filter/app.vue
@@ -175,7 +175,6 @@ const extendDownload = computed(() => [
 ])
 
 function removeCOFromList(ids) {
-  L
   list.value = list.value.filter((item) => !ids.includes(item.id))
   selectedIds.value = selectedIds.value.filter((id) => !ids.includes(id))
 }