+++ app/javascript/vue/tasks/observation_matrices/matrix_column_coder/MatrixColumnCoder/MediaDescriptor/MediaDescriptor.vue
@@ -129,7 +129,7 @@ export default {
     return {
       filterList: [],
       dropzoneObservation: {
-        paramName: 'observation[images_attributes][][image_file]',
+        paramName: 'observation[depictions_attributes][][image_attributes][image_file]',
         url: '/observations',
         autoProcessQueue: true,
         headers: {